+++ src/library/pthreadwrappers.cpp
@@ -27,6 +27,7 @@
 #include "backtrace.h"
 #include "hook.h"
 #include "timewrappers.h" // gettimeofday()
+#include "GameHacks.h"
 
 #include <errno.h>
 #include <unistd.h>
@@ -733,6 +734,14 @@ int pthread_setname_np (const char *name)
         GlobalState::setNoLog(true);
     }
 
+    if (strcmp(name, ".NET Finalizer") == 0) {
+#ifdef __unix__
+        GameHacks::setFinalizerThread(ThreadManager::getThreadTid(target_thread));
+#elif defined(__APPLE__) && defined(__MACH__)
+        GameHacks::setFinalizerThread(ThreadManager::getThreadTid());
+#endif
+    }
+    
     if (shared_config.game_specific_sync & SharedConfig::GC_SYNC_CELESTE) {
         if ((strcmp(name, "OVERWORLD_LOADE") == 0) ||
             (strcmp(name, "LEVEL_LOADER") == 0) ||